refactor: importgroups refactored
