+++ analyzer/model/tabledriveninfo.go
@@ -1,0 +1,132 @@
+package model
+
+import "go/ast"
+
+// TableDrivenInfo contains information about table-driven test.
+type TableDrivenInfo struct {
+	// Range that iterates over the tests and call t.Run
+	Range *ast.RangeStmt
+	// FormatType is either "map" or "slice".
+	FormatType string
+	// Inlined is true if the table is declared in the range statement.
+	Inlined bool
+	// Block is the body of the t.Run function.
+	Block *ast.BlockStmt
+}
+
+// newTableDrivenInfo returns information about a table driven test or nil if it's not a table-driven test.
+//
+//nolint:gocognit,funlen // refactor later
+func newTableDrivenInfo(testVar string, funcDecl *ast.FuncDecl) *TableDrivenInfo {
+	var stmts []ast.Stmt
+	if funcDecl.Body != nil {
+		stmts = funcDecl.Body.List
+	}
+
+	identifiers := make(map[string]*ast.CompositeLit)
+
+	var rangeStmt *ast.RangeStmt
+
+	for _, stmt := range stmts {
+		switch node := stmt.(type) {
+		// possible identifiers that can be used in a table-driven test
+		case *ast.AssignStmt:
+			if len(node.Rhs) != 1 {
+				continue
+			}
+
+			mapOrSliceCompositeLit := isMapOrSliceCompositeLit(node.Rhs[0])
+			if mapOrSliceCompositeLit == nil {
+				continue
+			}
+
+			if len(node.Lhs) != 1 {
+				continue
+			}
+
+			if ident, ok := node.Lhs[0].(*ast.Ident); ok {
+				identifiers[ident.Name] = mapOrSliceCompositeLit
+			}
+		// possible for loops that can be used in a table-driven test
+		case *ast.RangeStmt:
+			// the next instruction in a range stmt needs to be a t.Run
+			if node.Body != nil && len(node.Body.List) != 1 {
+				continue
+			}
+
+			exprStmt, isExprStmt := node.Body.List[0].(*ast.ExprStmt)
+			if !isExprStmt {
+				continue
+			}
+
+			callExpr, isCallExpr := exprStmt.X.(*ast.CallExpr)
+			if !isCallExpr {
+				continue
+			}
+
+			selectorExpr, isSelectorExpr := callExpr.Fun.(*ast.SelectorExpr)
+			if !isSelectorExpr {
+				continue
+			}
+
+			//nolint:lll // long line
+			if ident, isIdent := selectorExpr.X.(*ast.Ident); !isIdent || ident.Name != testVar || selectorExpr.Sel.Name != "Run" {
+				continue
+			}
+
+			funcLit, isFuncLit := callExpr.Args[1].(*ast.FuncLit)
+			if !isFuncLit {
+				continue
+			}
+			// from here, it's a table-driven test, we need to check whether is map/slice or inlined
+			rangeStmt = node
+
+			switch n := node.X.(type) {
+			case *ast.Ident:
+				// identifier must be declared before and be used as range
+				if _, isDeclaredBefore := identifiers[n.Name]; !isDeclaredBefore {
+					continue
+				}
+
+				// returns the second parameter of t.Run with the function
+				if len(callExpr.Args) != 2 {
+					continue
+				}
+
+				// is non-inlined and the param contains whether is map/slice
+				formatType := "map"
+				if _, isSlice := identifiers[n.Name].Type.(*ast.ArrayType); isSlice {
+					formatType = "slice"
+				}
+
+				return &TableDrivenInfo{
+					Range:      rangeStmt,
+					FormatType: formatType,
+					Inlined:    false,
+					Block:      funcLit.Body,
+				}
+			case *ast.CompositeLit:
+				// is inlined
+				if isMapOrSliceCompositeLit(n) == nil {
+					continue
+				}
+
+				formatType := "map"
+				if _, isSlice := n.Type.(*ast.ArrayType); isSlice {
+					formatType = "slice"
+				}
+
+				return &TableDrivenInfo{
+					Range:      rangeStmt,
+					FormatType: formatType,
+					Inlined:    true,
+					Block:      funcLit.Body,
+				}
+			default:
+				continue
+			}
+		}
+	}
+
+	return nil
+}
